update: restructure register page
