+++ frontend/src/App.js
@@ -6,7 +6,7 @@ import React from 'react';
 import './App.css';
 import Landing from './pages/shared/Landing/Landing';
 import Login from './pages/shared/Login/Login';
-import RegisterPage from './pages/RegisterPage';
+import Register from './pages/shared/Register/Register';
 import DashboardPage from './pages/DashboardPage';
 import ProfilePage from './pages/ProfilePage';
 import UpdateProfilePage from './pages/UpdateProfilePage';
@@ -26,7 +26,7 @@ function App() {
         <Switch>
           <Route path="/" exact component={Landing} />
           <Route path="/login" exact component={Login} />
-          <Route path="/register" exact component={RegisterPage} />
+          <Route path="/register" exact component={Register} />
           <Route path="/dashboard" exact component={DashboardPage} />
           <Route path="/profile" exact component={ProfilePage} />
           <Route path="/profile/update" exact component={UpdateProfilePage} />
