+++ frontend/src/api/StudentAPI.js
@@ -13,7 +13,26 @@ const StudentAPI = {
     };
   
     return fetch(`${process.env.REACT_APP_API}/login`, options);
-  }
+  },
+
+  register: (params) => {
+    const options = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'X-Requested-With': 'XMLHttpRequest'
+      },
+      body: JSON.stringify({
+        name: params.name,
+        username: params.username,
+        email: params.email,
+        password: params.password,
+        password_confirmation: params.passwordConfirm
+      })
+    };
+
+    return fetch(`${process.env.REACT_APP_API}/register`, options);
+  },
 };
 
 export default StudentAPI;
