+++ frontend/src/helpers/UserAuthApi.js
@@ -1,23 +1,4 @@
 const UserAuthApi = {
-  register: (params) => {
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Requested-With': 'XMLHttpRequest'
-      },
-      body: JSON.stringify({
-        name: params.name,
-        username: params.username,
-        email: params.email,
-        password: params.password,
-        password_confirmation: params.passwordConfirm
-      })
-    };
-
-    return fetch(`${process.env.REACT_APP_API}/register`, options);
-  },
-
   logout: (params) => {
     const options = {
       method: 'POST',
