+++ frontend/src/shared/Footer/Footer.js
@@ -2,7 +2,7 @@ const Footer = (props) => {
   return (
     <>
       <footer className={props.footerClass}>
-        <p>All Rights Reserved. &copy; Copyright 2021</p>
+        <p>E-Learning System. &copy; Copyright 2021</p>
       </footer>
     </>
   );
