+++ frontend/src/shared/FormTemplate/FormTemplate.module.css
@@ -1,0 +1,3 @@
+.fontWeight700 {
+  font-weight: 700;
+}
